refactor(shared): 홈 CTA 배너를 공통 컴포넌트로 추출해 마이홈과 공유 (#243)
