+++ src/shared/ui/index.ts
@@ -5,6 +5,7 @@ export * from './Avatar'
 export * from './Breadcrumb'
 export * from './AvatarGroup'
 export * from './Container'
+export * from './CtaBanner'
 export * from './DetailLink'
 export * from './Badge'
 export * from './FilterChip'
